pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctf-copilot"
version = "0.1.0"
description = "A real-time AI-assisted CTF copilot — your senior pentester in the terminal"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "pyyaml>=6.0",
    "anthropic>=0.25",
]

[project.scripts]
ctf     = "ctf_copilot.cli:cli"
ctf-log  = "ctf_copilot.commands.log_cmd:main"
ctf-wrap = "ctf_copilot.commands.wrap_cmd:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]
# Optional: Vector RAG knowledge base (past writeup similarity search)
# Install with: pip install ctf-copilot[rag]
rag = [
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ctf_copilot*"]

[tool.setuptools.package-data]
"ctf_copilot" = ["rules/*.yaml"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["ctf_copilot"]
omit   = ["tests/*", "*/migrations/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
